+++ src/bundles-processor.test.ts
@@ -1,33 +1,33 @@
 import { BundlesProcessor } from "./bundles-processor";
 import test, { GenericTestContext, Context } from "ava";
 import Vinyl from "vinyl";
-import { VinylExtension, BundlerStreamFactory } from "./main";
+import { BundlerStreamFactory } from "./main";
 import { Transform, TransformCallback, Readable } from "stream";
 
 test("BundlesProcessor with iterable inputs empty", async t => {
-    const files: Map<string, (Vinyl & VinylExtension)> = new Map();
+    const files: Map<string, [Vinyl, number]> = new Map();
     const bundles: Map<string, string[]> = new Map();
 
     TestBundler(t, [[], new Map()], await BundlesProcessor(files, bundles, BundleStreamFactory));
-    
+
 });
 
 test("BundlesProcessor with files but no bundles", async t => {
-    const files: Map<string, (Vinyl & VinylExtension)> = new Map();
-    files.set("test", MakeExtendedVinyl("test", "test"));
+    const files: Map<string, [Vinyl, number]> = new Map();
+    files.set("test", [MakeVinyl("test", "test"), 0]);
     const bundles: Map<string, string[]> = new Map();
 
     TestBundler(t, [[], new Map()], await BundlesProcessor(files, bundles, BundleStreamFactory));
 });
 
 test("BundlesProcessor with files and bundles", async t => {
-    const files: Map<string, (Vinyl & VinylExtension)> = new Map();
-    files.set("test", MakeExtendedVinyl("test", "test"));
+    const files: Map<string, [Vinyl, number]> = new Map();
+    files.set("test", [MakeVinyl("test", "test"), 0]);
     const bundles: Map<string, string[]> = new Map();
     bundles.set("test", ["test"]);
 
     const resultChunks: any[] = [
-        MakeExtendedVinyl("test", "test")
+        MakeVinyl("test", "test")
     ];
     const resultPaths: Map<string, string[]> = new Map();
     resultPaths.set("test", ["test"]);
@@ -42,7 +42,7 @@ function TestBundler(t: GenericTestContext<Context<any>>, expected: [any[], Map<
     // Sort result paths
     expected[1].forEach(paths => paths.sort());
     actual[1].forEach(paths => paths.sort());
-    
+
     // Check result paths
     t.deepEqual(expected[1], actual[1]);
 }
@@ -66,17 +66,12 @@ class TestTransform extends Transform {
  * Makes an extended Vinyl object. For testing purposes.
  * @param contents Contents used to make buffer.
  * @param path Path of file.
- * @param precedence Precedence value. Optional.
  */
-function MakeExtendedVinyl(contents: string, path: string, precedence: number = 0): (Vinyl & VinylExtension) {
-    return Object.assign(
-        new Vinyl({
-            contents: Buffer.from(contents),
-            path
-        }), {
-            Precedence: precedence
-        }
-    );
+function MakeVinyl(contents: string, path: string): Vinyl {
+    return new Vinyl({
+        contents: Buffer.from(contents),
+        path
+    });
 }
 
 const BundleStreamFactory: BundlerStreamFactory = (src: Readable): Transform => {